moved provider level up, added pages component, moved theme fetch to app
