+++ packages/client/src/components/pages/index.ts
@@ -1,0 +1,1 @@
+export { default as Pages } from './Pages'
