+++ packages/client/src/modules/entranceForm/entranceForm.module.scss
@@ -20,7 +20,7 @@
 
     &:hover {
       background-color: var(--primary);
-      color: var(--background-color);
+      color: var(--background-color-dark);
     }
   }
