+++ packages/client/src/pages/main/modules/profile/profile.module.scss
@@ -83,7 +83,7 @@
   color: var(--primary);
   &:hover {
     background-color: var(--primary);
-    color: var(--secondary)
+    color: var(--background-color-dark)
   }
 }
